requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
bitsandbytes==0.47.0
certifi==2025.8.3
charset-normalizer==3.4.3
cut-cross-entropy==25.1.1
datasets==3.6.0
diffusers==0.35.1
dill==0.3.8
docstring-parser==0.17.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.3.0
hf-transfer==0.1.9
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
importlib-metadata==8.7.0
jinja2==3.1.6
markdown-it-py==4.0.0
markupsafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
msgspec==0.19.0
multidict==6.6.4
multiprocess==0.70.16
networkx==3.5
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pandas==2.3.2
peft==0.17.1
pillow==11.3.0
propcache==0.3.2
protobuf==6.32.1
psutil==7.0.0
pyarrow==21.0.0
pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
safetensors==0.6.2
sentencepiece==0.2.1
setuptools==80.9.0
shtab==1.7.2
six==1.17.0
sympy==1.14.0
tokenizers==0.22.0
torch==2.8.0
torchao==0.13.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.1
triton==3.4.0
trl==0.23.0
typeguard==4.4.4
typing-extensions==4.15.0
tyro==0.9.31
tzdata==2025.2
unsloth==2025.9.4
unsloth-zoo==2025.9.5
urllib3==2.5.0
wheel==0.45.1
xformers==0.0.32.post2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
